Pasando recursos a MoreDetailsView
