+++ AK Apolo/app/src/main/java/ayds/apolo/songinfo/moredetails/fulllogic/controller/MoreDetailsControllerModule.kt
@@ -1,14 +1,11 @@
 package ayds.apolo.songinfo.moredetails.fulllogic.controller
 
-import ayds.apolo.songinfo.home.controller.HomeControllerImpl
-import ayds.apolo.songinfo.home.model.HomeModelModule
-import ayds.apolo.songinfo.home.view.HomeView
 import ayds.apolo.songinfo.moredetails.fulllogic.model.MoreDetailsModelModule
-import ayds.apolo.songinfo.moredetails.fulllogic.view.MoreDetailsView
+import ayds.apolo.songinfo.moredetails.fulllogic.view.MoreDetailsViewActivity
 
 object MoreDetailsControllerModule {
 
-    fun viewStart(moreDetailsView: MoreDetailsView) {
+    fun viewStart(moreDetailsView: MoreDetailsViewActivity) {
         MoreDetailsControllerImpl(MoreDetailsModelModule.getMoreDetailsModel()).apply {
             setMoreDetailsView(moreDetailsView)
         }
