+++ AK Apolo/app/src/main/java/ayds/apolo/songinfo/moredetails/fulllogic/view/MoreDetailsUiEvent.kt
@@ -1,0 +1,4 @@
+package ayds.apolo.songinfo.moredetails.fulllogic.view
+
+class MoreDetailsUiEvent {
+}
